Move pbf-parser-stuff into one file to work on it
