--- src/osm/pbf/mod.rs
@@ -1,5 +1,0 @@
-mod reader;
-pub use reader::Reader;
-
-mod parser;
-pub use parser::Parser;
